Move most of the distributed code into an extension.
