+++ Project.toml
@@ -5,15 +5,16 @@ version = "0.3.8"
 [deps]
 EnumX = "4e289a0a-7415-4d19-859d-a7e5c4648b56"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
-MPIPreferences = "3da0fdf6-3ccc-4f1b-acd9-58baa6c99267"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
 NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
-PartitionedArrays = "5a9dfac6-5c52-46f7-8278-5e2210713be9"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Tensors = "48a634ad-e948-5137-8d70-aa71f2a747f4"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
+
+[weakdeps]
+MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
+PartitionedArrays = "5a9dfac6-5c52-46f7-8278-5e2210713be9"
 
 [compat]
 EnumX = "1"
@@ -31,7 +32,9 @@ FerriteGmsh = "4f95f4f8-b27c-4ae5-9a39-ea55e634e36b"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 Gmsh = "705231aa-382f-11e9-3f0c-b7cb4346fdeb"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
+MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 NBInclude = "0db19996-df87-5ea3-a455-e3a50d440464"
+PartitionedArrays = "5a9dfac6-5c52-46f7-8278-5e2210713be9"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
